change default handling of nonbooleans - always trying two possible values (0 and 1 or "" and "nonempty")
this exposed a bug in how conf preserves invalid configurations when an
int is required but non is provided.
in any configuration mode anything but an integer is rejected, but
conf would not complain on prior configurations where the integer is
not set.

this breaks our differential testing mechanisms, as ground truth cannot
be established with `conf` for those cases.

added a failing test case to illustrate the point